Moved interpolate to separate file
